update DEM_Tile_Example to use the max zoom level for resolution calc (#676)
When zooming in more than a certain level, the UDF used to throw a
`ValueError: Integers to negative integer powers are not allowed.`.

Since the default zoom value used in the integration tests is greater
than the threshold used here, it always used to fail. This PR should fix
that